refacotr: Domain 검증코드 도메인 클래스 내 메서드화
